Move statuscodes to parcels root
