+++ parcels/__init__.py
@@ -23,6 +23,7 @@ from parcels.particlefile import ParticleFile
 from parcels.particleset import ParticleSet
 from parcels._logger import logger
 from parcels.utils import *
+from parcels.statuscodes import StatusCode
 from parcels._tutorial import download_example_dataset, list_example_datasets
 
 _stdlib_warnings.warn(
