+++ parcels/particleset.py
@@ -16,7 +16,7 @@ from parcels.kernels import AdvectionRK4
 from parcels.particle import KernelParticle, Particle, create_particle_data
 from parcels.utils.converters import convert_to_flat_array
 from parcels._logger import logger
-from parcels.utils.statuscodes import StatusCode
+from parcels.statuscodes import StatusCode
 from parcels._warnings import ParticleSetWarning
 
 __all__ = ["ParticleSet"]
