+++ parcels/utils/__init__.py
@@ -1,3 +1,2 @@
 from .converters import *
 from .interpolation_utils import *
-from .statuscodes import *
