+++ tests/test_interpolation.py
@@ -19,7 +19,7 @@ from parcels.interpolators import (
 from parcels.particle import Particle, Variable
 from parcels.particlefile import ParticleFile
 from parcels.particleset import ParticleSet
-from parcels.utils.statuscodes import StatusCode
+from parcels.statuscodes import StatusCode
 from parcels.uxgrid import UxGrid
 from parcels.xgrid import XGrid
 from tests.utils import TEST_DATA
