+++ tests/test_particleset_execute.py
@@ -19,7 +19,7 @@ from parcels._datasets.structured.generated import simple_UV_dataset
 from parcels._datasets.structured.generic import datasets as datasets_structured
 from parcels._datasets.unstructured.generic import datasets as datasets_unstructured
 from parcels.particlefile import ParticleFile
-from parcels.utils.statuscodes import FieldInterpolationError, FieldOutOfBoundError, TimeExtrapolationError
+from parcels.statuscodes import FieldInterpolationError, FieldOutOfBoundError, TimeExtrapolationError
 from parcels.uxgrid import UxGrid
 from parcels.xgrid import XGrid
 from tests import utils
